Move palette construction to PalettePicker class
